Cached IDs should be strings
Previously, when using pickle, we could preserve cache keys as
ints. However, we stopped using pickle due to security concerns,
and JSON does not allow keys to be anything but strings.
This changes the structure of internal cache from int:str to
str:str, which can cause issues when Uyuni expects int and
spacecmd sends str.

This change fixes the issue, and casts IDs from internal cache to
ints.